requirements.txt
xkcdpass
python-gnupg
